# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/article/timeline": {
      "get": {
        "operationId": "ArticleController_getTimeline",
        "parameters": [],
        "responses": {
          "200": {
            "description": "获取时间轴成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取时间轴",
        "tags": [
          "文章"
        ]
      }
    },
    "/api/article/prevNext/{id}": {
      "get": {
        "operationId": "ArticleController_getPrevNext",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "文章"
        ]
      }
    },
    "/api/article": {
      "post": {
        "operationId": "ArticleController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateArticleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "文章"
        ]
      },
      "get": {
        "operationId": "ArticleController_findAllByPages",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "category",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tag",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "title",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "文章"
        ]
      }
    },
    "/api/article/uploadTime": {
      "get": {
        "operationId": "ArticleController_getUploadTime",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "文章"
        ]
      }
    },
    "/api/article/{id}": {
      "get": {
        "operationId": "ArticleController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "文章"
        ]
      },
      "put": {
        "operationId": "ArticleController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateArticleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "文章"
        ]
      },
      "delete": {
        "operationId": "ArticleController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "文章"
        ]
      }
    },
...
  },
  "info": {
    "title": "朝阳",
    "description": "朝阳博客相关的Api",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateArticleDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateArticleDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}